backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
redis==4.6.0
django-redis==5.4.0
Pillow==10.0.1
django-extensions==3.2.3
python-decouple==3.8
celery==5.3.4
djangorestframework-simplejwt==5.3.0
django-filter==23.3
whitenoise==6.6.0